This is a lightweight refactor to clean up the root directory by organizing files into basic folders. The goal is to slightly improve structure and readability without introducing breaking changes or heavy rewrites.
